+++ 06-SSM-02-fp/src/main/java/com/fc/service/UserService.java
@@ -7,13 +7,13 @@ import java.util.ArrayList;
 
 @Service
 public interface UserService{
-    int del(long id);
+    int del(Integer id);
 
     int add(User user);
 
     int update(User user);
 
-    User findById(Long id);
+    User findById(Integer id);
 
     ArrayList<User> findAll(Integer pageNo, Integer pageSize);
 }
